Deleted tutorial background.js code
Added a line between the buttons and the list of definitions currently saved

changed back from <form> to just a <div> containing the buttons; forms send a 'submit' command whenever a button presses; in our case that just refreshes the page. Since it isn't really a form, a <div> is fine and works better.

Surrounded all JS code in an eventlistener for DOMContentLoadded, which will ensure none of our code is attempted to run before the webpage loads.

Added actual definition list capability; now shows which definitions are saved on the right side of the window (basically, what will be sent to [output] -- currently clipboard.